% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_chains.R
\name{plot_trace_graph}
\alias{plot_trace_graph}
\title{Plot a Derivation Graph}
\usage{
plot_trace_graph(trace_model, adam_derivations = NULL, ...)
}
\arguments{
\item{trace_model}{A \code{trace_model} from \code{\link[=build_trace_model]{build_trace_model()}}.}

\item{adam_derivations}{Optional ADaM-to-ADaM derivations, as in
\code{\link[=trace_chains]{trace_chains()}}.}

\item{...}{Passed to \code{\link[igraph:plot.igraph]{igraph::plot.igraph()}}.}
}
\value{
The \code{igraph} object, invisibly.
}
\description{
Draws the derivation graph behind \code{\link[=trace_chains]{trace_chains()}} using \code{igraph}: ADaM and
SDTM variables as nodes, derivation links as directed edges.
}
\examples{
\dontrun{
plot_trace_graph(tm, adam_derivations = deriv)
}

}
