% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_model.R
\name{print.trace_model}
\alias{print.trace_model}
\title{Print Trace Model}
\usage{
\method{print}{trace_model}(x, ...)
}
\arguments{
\item{x}{A \code{trace_model} object.}

\item{...}{Ignored.}
}
\description{
Print Trace Model
}
